Adapt deamons to walletd
As a first step towards isolating secrets in walletd, remove them from
the peerd keyopts. Peerd now takes the secret as an argument
when launched from farcasterd. Farcasterd gets the secrets from walletd
with the GetSecret rpc call.

In case the secrets have not been fetched yet before they are required
in farcasterd, e.g. to launch a peerd instance in listen mode, GetSecret
is called as well, with a context as an extra argument. The context
holds information to execute the original rpc again once the secrets are
received by farcasterd. In essence, farcasterd needs to loopback from
its Secret endpoint to the original rpc with the help of the provided
context. Since the handling of RPC calls is blocking, we can't call
farcasterd directly, so an additional Loopback endpoint is implemented
in walletd. Loopback just takes the context as a single argument, and
calls its respective RPC call in farcasterd.