Syncer functional test: Test sweep minimum amount
To test that the minimum amount is respected, first send an XMR amount
below to required minimum to the sweep address, then send another amount
to complete the required minimum. The syncer should only send a single
event for the sweep of the entire amount, and not two events. A second,
non-blocking receive on the zmq socket should therefore error.