Swapd: Support cancel swap
A swap should be cancelable by the user if its side has not locked any
coins yet.

For Alice, this means a swap can be canceled without additional handling
up until she reaches state RefundSigA.

For Bob, this means a swap can be canceled without additional handling as
long as it still is in its original start state. If it is in commit,
reveal, or corearb state and has not yet received the refund procedure
signatures from Alice, it additionally attempts to sweep any coins that
were sent to the funding address.