Add Support for electrum client Tor proxy
This adds support to route the traffic of the electrum client through a
Tor proxy.

To simplify the argument parsing and retain compatibility with
positional arugments, the default flag value for the socks proxy was
removed. However, this functionality was never properly implemented in
the first place. The socks proxy address always had to be passed in
additionally, the default value was never used.