Syncer: Remove polling switch
This removes the polling switch implementation. The polling is never
used in the Monero syncer and is used in a very wasteful fashion on the
Bitcoin side.

With this commit, the Bitcoin syncer will only support subscribing to
height and script events and not continuously polling for their change.
This simplifies the syncer code a bit and will make future refactoring
easier.