Farcasterd: Retrieve keys on startup
The changeset in this PR removes the pending requests in farcasterd
currently used for retrieving the peerd keys and instead caches the keys
after they are retrieved from walletd on startup. While the keys have
not been retrieved yet, or the required services are not up yet make and
take swap will fail and report the failure to the cli.

I realize a lot of engineering effort went into ensuring that the
service is addressable immediately and we don't have these kind of
checks, but hear me out here.

In order to support Tor for makers / listeners in a user friendly
manner, farcasterd has to create a Tor hidden service once at startup.
Additionally, this would involve moving the bind and public addresses
from the make command to the arguments used by farcaster on startup to
ensure the user does not leak his connection by accidentally binding on
a non-Tor Ip. Since this move would include waiting for the Tor hidden
service to start up as well, I was thinking that adding this short
period on startup where the required data has to be retrieved is
acceptable, especially since there are some assumptions about the
service running already.

Accepting this commit will also mean that there eventually will only be
single static bind address that can be used at a time.