Farcaster: Implement much of farcasterd's logic as state machines
This commit re-writes the runtime of the farcaster daemon as a
collection of deterministic finite automata. It defines traits and types
for a state machine that consumes events and manipulates a runtime. The
actual automata are documented and described in their respective
implementations.

Requests pertaining to offers and swaps are processed in the
TradeStateMachine. It tracks the lifetime of a single offer and swap
starting at either a make offer, take offer or a swap restore. All
running trade state machines are held in a collection in the runtime.

Similarly, syncer related requests are processed by the
SyncerStateMachine and held in a separate collection in the runtime.

A function pre-processes the requests by selecting the correct state
machine and creating an event to process the request.

A number of bugs were discovered and fixed in the process of this
re-write. The syncer launching and tracking was still not properly done
for swap restore and sweep requests. Some changes in the state were not
executed in the right order or after the right request. This led to
incorrect reporting of open offers, available services, running
swaps, and incorrect handling of existing connections.