Farcasterd: Refactor listen and connect
This is done with the goal of making the processing of make and take
offer easier to read by removing most of the and_then statements and
unnesting some of the logic.

Coincidentally, this also solves a bug where the peer services are not
removed from the spawning_services HashSet and where the connect
function might double-spawn a connection that has not said Hello to
farcasterd yet.