Cli: Implement endpoint for checking syncer health
This adds a cli command "health-check" for checking the configuration
and required environment (e.g. the to-be-connected coin daemons,
wallets, etc.) of a syncer.

To this end, this commit also adds a HealthCheck event that can be
dispatched to a syncer. The various syncers make simple calls to the
respective coin daemons and report their success back in the form of a
HealthResult event.

The "health-check" command makes use of the syncer state machine in
farcasterd to startup syncers and route calls back to the client.