Peerd: Use message receipts to confirm receival
Peer messages sent to a remote peer are now cached until the remote peer
has replied with a corresponding receipt.

If the connection fails in a listener spawned peerd, the cache is
emptied before peerd shuts down and the messages are sent back to their
respective origin (swapd). Swapd then processes them into its own cache.

If the connection fails in a connecting peerd, the cache is only emptied
once the reconnect has been a success. Messages are not sent back to
their origin.